Started a new chess engine project.
I created a file for a bitboard type and it's
associated functions. I also created a struct that will represent
all the state needed in a chess game and it's associated functions.
Finally, I created a FEN parser, which parses FEN strings into
piece positions and returns them in a game state we can use.