Added interactive mode, refactored 'engine' code
put most non-trivial commands into their own static functions,
moved perft to it's own file.

added "interactive mode" which allows the engine, when running in a
tty, to act more as a repl I can use to test instead of strict uci
compliance. For example, when using the position command, automatically
print the new position to the console.